refactor(http): hyper backend
